Refactor java code to change all id types from Long to Integer + a bunch of reformatting

+++ src/main/java/amt/dto/SampleDTO.java
@@ -3,9 +3,9 @@ package amt.dto;
 import java.time.LocalDateTime;
 
 public record SampleDTO(
-        Long id,
+        Integer id,
         String name,
         String filepath,
         Double duration,
-        LocalDateTime createdAt
-) {}
+        LocalDateTime createdAt) {
+}
